👷 Rewrite reaction to comment action using `github-script`
